api/requirements.txt
aiosqlite==0.19.0
anyio==3.7.0
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.9
PyJWT==2.7.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
websockets==11.0.3
